More cleaning

+++ lib/goom/src/gfontrle.h
@@ -1,14 +1,17 @@
 #ifndef _GFONTRLE_H
 #define _GFONTRLE_H
 
+#include <cstdint>
+
 struct TheFont
 {
-  unsigned int width;
-  unsigned int height;
-  unsigned int bytes_per_pixel;
-  unsigned int rle_size;
+  uint16_t width;
+  uint16_t height;
+  uint16_t bytes_per_pixel;
+  uint16_t rle_size;
   unsigned char rle_pixel[49725];
 };
+
 extern const TheFont the_font;
 
 #endif
